- iPhone OS port - Re-Synced GLESRenderSystem with GLRenderSystem and added some more extension detection - Added support for ARM architecture and CPU feature detection - Eliminated deprecation warnings when targeting Mac OS X 10.5/10.6 - warnings-- 338 vs. 56, 83% reduction in full debug build(all plugins, samples, tests) with GCC on Mac OS X - Consolidated macPlugins and macBundlePath into new macUtils file - Remove duplicated macBundlePath code in ExampleApplication.h - Fixed OS X TODO item to use dlopen/dlsym/dlclose to load plugins. - CMake 	- Fleshed out OpenGL ES detection and CMake support.  Likely not done but more improved. 	- Some placeholder stuff for an OpenGL ES 2.0 rendersystem in CMakefiles 	- iPhone OS support 	- Moved code and resources from the Mac directory to OgreMain 	- Building Tests cannot find config.h on OS X

+++ PlugIns/OctreeSceneManager/src/OgreTerrainPage.cpp
@@ -72,9 +72,9 @@ namespace Ogre {
     {
         //setup the neighbor links.
 
-        for ( size_t j = 0; j < tilesPerPage; j++ )
+        for ( ushort j = 0; j < tilesPerPage; j++ )
         {
-            for ( size_t i = 0; i < tilesPerPage; i++ )
+            for ( ushort i = 0; i < tilesPerPage; i++ )
             {
                 if ( j != tilesPerPage - 1 )
                 {
@@ -121,9 +121,9 @@ namespace Ogre {
 	//-------------------------------------------------------------------------
 	void TerrainPage::setRenderQueue(uint8 qid)
 	{
-		for ( size_t j = 0; j < tilesPerPage; j++ )
+		for ( ushort j = 0; j < tilesPerPage; j++ )
 		{
-			for ( size_t i = 0; i < tilesPerPage; i++ )
+			for ( ushort i = 0; i < tilesPerPage; i++ )
 			{
 				if ( j != tilesPerPage - 1 )
 				{
